fix:Alterando os tipos dos parametros do saque e depósito

+++ POO_projeto/src/model/ContaCorrente.java
@@ -48,7 +48,7 @@ public class ContaCorrente implements IConta {
 	}
 
 	@Override
-	public void depositar(int quantia) {
+	public void depositar(float quantia) {
 		if (status) {
 			if(quantia >0) {
 				this.saldo+=quantia;
@@ -103,7 +103,7 @@ public class ContaCorrente implements IConta {
 	}
 
 	@Override
-	public void sacar(int quantia) {
+	public void sacar(float quantia) {
 		if(quantia+TAXA_TRANSACAO_CORRENTE_SAQUE>this.saldo) {
 		this.saldo -=quantia;
 		float valor = (float)quantia;
